requirements.txt
shyaml
pyyaml
envbash
munch
GitPython
pyfiglet